Move non legacy stt models out from legacy module
